+++ OthelloSolverCommandLine/InternalRoutines.h
@@ -2,3 +2,6 @@
 void SetFullDirPathForRun(const char* outputDir, int boardSize);
 char* GetFullDirPathForRun();
 bool CreateFullPathForRun(const char* outputDir, int boardSize);
+char *GetFullFilePathBaseNameForBoardLevel(int boardLevel);
+char *GetFullFilePathBaseNameForMoveLevel(int boardLevel);
+